started working on the new theming

+++ src/main/java/de/aivot/GoverBackend/theme/services/ThemeService.java
@@ -1,18 +1,18 @@
 package de.aivot.GoverBackend.theme.services;
 
+import de.aivot.GoverBackend.department.filters.DepartmentFilter;
+import de.aivot.GoverBackend.department.repositories.DepartmentRepository;
 import de.aivot.GoverBackend.form.filters.FormVersionFilter;
-import de.aivot.GoverBackend.form.repositories.FormRepository;
 import de.aivot.GoverBackend.form.repositories.FormVersionRepository;
 import de.aivot.GoverBackend.lib.exceptions.ResponseException;
 import de.aivot.GoverBackend.lib.models.Filter;
 import de.aivot.GoverBackend.lib.services.EntityService;
-import de.aivot.GoverBackend.theme.entities.Theme;
+import de.aivot.GoverBackend.theme.entities.ThemeEntity;
 import de.aivot.GoverBackend.theme.repositories.ThemeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Nonnull;
@@ -20,37 +20,45 @@ import javax.annotation.Nullable;
 import java.util.Optional;
 
 @Service
-public class ThemeService implements EntityService<Theme, Integer> {
+public class ThemeService implements EntityService<ThemeEntity, Integer> {
     private final ThemeRepository themeRepository;
-    private final FormRepository formRepository;
     private final FormVersionRepository formVersionRepository;
+    private final DepartmentRepository departmentRepository;
 
     @Autowired
-    public ThemeService(
-            ThemeRepository themeRepository,
-            FormRepository formRepository,
-            FormVersionRepository formVersionRepository) {
+    public ThemeService(ThemeRepository themeRepository,
+                        FormVersionRepository formVersionRepository,
+                        DepartmentRepository departmentRepository) {
         this.themeRepository = themeRepository;
-        this.formRepository = formRepository;
         this.formVersionRepository = formVersionRepository;
+        this.departmentRepository = departmentRepository;
     }
 
     @Nonnull
     @Override
-    public Theme create(@Nonnull Theme entity) throws ResponseException {
+    public ThemeEntity create(@Nonnull ThemeEntity entity) throws ResponseException {
         entity.setId(null);
         return themeRepository.save(entity);
     }
 
     @Override
-    public void performDelete(@Nonnull Theme entity) throws ResponseException {
+    public void performDelete(@Nonnull ThemeEntity entity) throws ResponseException {
         var formSpec = FormVersionFilter
                 .create()
                 .setThemeId(entity.getId())
                 .build();
 
         if (formVersionRepository.exists(formSpec)) {
-            throw new ResponseException(HttpStatus.CONFLICT, "Das Farbschema wird noch von einem oder mehreren Formularen verwendet.");
+            throw ResponseException.conflict("Das Farbschema wird noch von einem oder mehreren Formularen verwendet.");
+        }
+
+        var depSpec = DepartmentFilter
+                .create()
+                .setThemeId(entity.getId())
+                .build();
+
+        if (departmentRepository.exists(depSpec)) {
+            throw ResponseException.conflict("Das Farbschema wird noch von einer oder mehreren Fachbereichen verwendet.");
         }
 
         themeRepository.delete(entity);
@@ -58,13 +66,13 @@ public class ThemeService implements EntityService<Theme, Integer> {
 
     @Nonnull
     @Override
-    public Page<Theme> performList(@Nonnull Pageable pageable, @Nullable Specification<Theme> specification, Filter<Theme> filter) {
+    public Page<ThemeEntity> performList(@Nonnull Pageable pageable, @Nullable Specification<ThemeEntity> specification, Filter<ThemeEntity> filter) {
         return themeRepository.findAll(specification, pageable);
     }
 
     @Nonnull
     @Override
-    public Theme performUpdate(@Nonnull Integer id, @Nonnull Theme entity, @Nonnull Theme existingEntity) throws ResponseException {
+    public ThemeEntity performUpdate(@Nonnull Integer id, @Nonnull ThemeEntity entity, @Nonnull ThemeEntity existingEntity) throws ResponseException {
         existingEntity.setName(entity.getName());
         existingEntity.setMain(entity.getMain());
         existingEntity.setMainDark(entity.getMainDark());
@@ -73,19 +81,21 @@ public class ThemeService implements EntityService<Theme, Integer> {
         existingEntity.setWarning(entity.getWarning());
         existingEntity.setInfo(entity.getInfo());
         existingEntity.setSuccess(entity.getSuccess());
+        existingEntity.setLogoKey(entity.getLogoKey());
+        existingEntity.setFaviconKey(entity.getFaviconKey());
 
         return themeRepository.save(existingEntity);
     }
 
     @Nonnull
     @Override
-    public Optional<Theme> retrieve(@Nonnull Integer id) {
+    public Optional<ThemeEntity> retrieve(@Nonnull Integer id) {
         return themeRepository.findById(id);
     }
 
     @Nonnull
     @Override
-    public Optional<Theme> retrieve(@Nonnull Specification<Theme> specification) {
+    public Optional<ThemeEntity> retrieve(@Nonnull Specification<ThemeEntity> specification) {
         return themeRepository.findOne(specification);
     }
 
@@ -95,7 +105,7 @@ public class ThemeService implements EntityService<Theme, Integer> {
     }
 
     @Override
-    public boolean exists(@Nonnull Specification<Theme> specification) {
+    public boolean exists(@Nonnull Specification<ThemeEntity> specification) {
         return themeRepository.exists(specification);
     }
 }
